Add support for suspending the zone
Next changes are done for adding support for suspending the zone:

* Added a new column ('status') in the table zone.

* Added API method SetZoneStatus that changes the zone status. This
method has 2 string arguments zonename and zonestatus. The zonename
argument is representing the zone that is being updated, and the
zonestatus argument ('active', 'suspended') is representing a new
zone status.

* Added API method GetZoneStatusBulk that will get the status of the
zones. The input parameter is an array of zone names, and the output
is an array of objects (zonename, zonestatus).

Resolves PROD-2913.

ChangeLog:

* [IMPROVEMENT] Added support for suspending the zone.